회원 상세

+++ src/main/resources/static/js/post.js
@@ -19,7 +19,12 @@ let index = {
 		});//blur
 		$("#btn-reply-save").on("click", () => {
 			this.replySave();
-		})
+		});//click
+		$("#btn-delete").on("click", () => {
+			if (confirm("포스트를 삭제하시겠습니까?")) {
+				this.delete();
+			}
+		});//click
 
 	},//init	
 	isEnabled: function() {
@@ -53,7 +58,7 @@ let index = {
 			videoId: this.extractVideoID($("#videoId").val())
 		};
 		$.ajax({
-			url: "/posts",
+			url: "/api/posts",
 			data: JSON.stringify(data),
 			type: "POST",
 			contentType: "application/json; charser=UTF-8",
@@ -70,27 +75,28 @@ let index = {
 		});
 	},//save
 	update: function() {
+		let id= $("#postId").val();
 		let data = {
-			id: $("#id").val(),
-			username: $("#username").val(),
-			password: $("#password").val(),
-			email: $("#email").val()
+			singer: $("#singer").val(),
+			title: $("#title").val(),
+			description: $("#description").val(),
+			videoId: this.extractVideoID($("#videoId").val())
 		};
 		$.ajax({
-			url: "/users",
+			url: `/api/posts/${id}`,
 			data: JSON.stringify(data),
 			type: "PUT",
 			contentType: "application/json; charser=UTF-8",
 			dataType: "JSON"
 		}).done(function(resp) {
 			if (resp.data == 1) {
-				alert("회원수정이 완료되었습니다.");
-				location.href = "/";
+				alert("포스트 수정이 완료되었습니다.");
+				location.href = `/posts/${id}`;
 			} else {
-				alert("회원수정을 실패하였습니다.");
+				alert("포스트 수정을 실패하였습니다.");
 			}
 		}).fail(function(error) {
-			alert("회원수정 중 문제가 발생했습니다. 잠시 후 다시 시도해주세요");
+			alert("포스트 수정 중 문제가 발생했습니다. 잠시 후 다시 시도해주세요");
 		});
 	},//update
 	extractVideoID: function(url) {
@@ -111,7 +117,7 @@ let index = {
 			postId: $("#postId").val()
 		};
 		$.ajax({
-			url: `/posts/${data.postId}/replys`,
+			url: `/api/posts/${data.postId}/replys`,
 			data: JSON.stringify(data),
 			type: "POST",
 			contentType: "application/json; charser=UTF-8",
@@ -127,28 +133,48 @@ let index = {
 		}).fail(function(error) {
 			alert("댓글등록 중 문제가 발생했습니다. 잠시 후 다시 시도해주세요");
 		});
-	}//replySave
+	},//replySave
+
+	delete: function() {
+		let postId = $("#postId").val();
+		$.ajax({
+			url: `/api/posts/${postId}`,
+			type: "DELETE",
+			contentType: "application/json; charser=UTF-8",
+			dataType: "JSON"
+		}).done(function(resp) {
+			if (resp.data == 1) {
+				alert("포스트 삭제가 완료되었습니다. 메인 페이지로 이동합니다.");
+				location.href="/";
+			} else {
+				alert("포스트 삭제를 실패하였습니다.");
+			}
+		}).fail(function(error) {
+			alert("삭제 중 문제가 발생했습니다. 잠시 후 다시 시도해주세요");
+		});
+	}
+
 }//index
 function getCommentList() {
 	//동영상을 끊기지 않고 댓글을 갱신하기 위해서 json 데이터를 받아 비동기 화면 전환을 구현
 	let = postId = $("#postId").val();
 	let userId = $("#userId").val();
 	$.ajax({
-		url: `/posts/${postId}/replys`,
+		url: `/api/posts/${postId}/replys`,
 		type: "GET",
 		contentType: "application/json; charser=UTF-8",
 		dataType: "JSON"
 	}).done(function(json) {
 		var html = "";
-		if(json==""){
-			html = "<font color='white'>댓글이 존재하지 않습니다.</font>";	
-		}else{
+		if (json == "") {
+			html = "<font color='white'>댓글이 존재하지 않습니다.</font>";
+		} else {
 			json.forEach(function(reply, idx) {
 				html += "<div class='d-flex py-2 text-white'><div class='flex-shrink-0'>";
 				html += "<img class='rounded-circle' src='../image/Profile-PNG-Clipart.png' style='width:50px;'>";
 				html += "</div><div class='ms-3'>";
 				html += "<div class='fw-bold'>" + reply.user.username + " &nbsp;<font size='2'>";
-				html += reply.createDate;
+				html += dateFormat(new Date(reply.createDate));
 				if (userId == reply.user.id) {
 					html += " <button class='btn btn-success btn-sm' onclick='replyDelete(" + reply.id + ")'>삭제</button>";
 				}//end if
@@ -161,10 +187,13 @@ function getCommentList() {
 	}).fail(function(error) {
 		alert(" 문제가 발생했습니다. 잠시 후 다시 시도해주세요");
 	});
-}
+}//getCommentList
 function replyDelete(replyId) {
+	if (!confirm("댓글을 삭제하시겠습니까?")) {
+		return;
+	}
 	$.ajax({
-		url: `/replys/${replyId}`,
+		url: `/api/replys/${replyId}`,
 		type: "DELETE",
 		contentType: "application/json; charser=UTF-8",
 		dataType: "JSON"
@@ -179,5 +208,19 @@ function replyDelete(replyId) {
 	}).fail(function(error) {
 		alert("댓글삭제 중 문제가 발생했습니다. 잠시 후 다시 시도해주세요");
 	});
-}
+}//replyDelete
+function dateFormat(date) {
+	//자바스크립트 날짜 포맷을 변경하기 위한 함수 선언
+        let month = date.getMonth() + 1;
+        let day = date.getDate();
+        let hour = date.getHours();
+        let minute = date.getMinutes();
+
+        month = month >= 10 ? month : '0' + month;
+        day = day >= 10 ? day : '0' + day;
+        hour = hour >= 10 ? hour : '0' + hour;
+        minute = minute >= 10 ? minute : '0' + minute;
+
+        return date.getFullYear() + '-' + month + '-' + day + ' ' + hour + ':' + minute;
+}//dateFormat
 index.init();
